[F] move isUrl to src/presenter/validation
